requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
boto3>=1.34.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
